chore: layout refac
