+++ src/lib/apis/auths/index.ts
@@ -1,5 +1,32 @@
 import { WEBUI_API_BASE_URL } from '$lib/constants';
 
+export const getSessionUser = async (token: string) => {
+	let error = null;
+
+	const res = await fetch(`${WEBUI_API_BASE_URL}/auths`, {
+		method: 'GET',
+		headers: {
+			'Content-Type': 'application/json',
+			Authorization: `Bearer ${token}`
+		}
+	})
+		.then(async (res) => {
+			if (!res.ok) throw await res.json();
+			return res.json();
+		})
+		.catch((err) => {
+			console.log(err);
+			error = err.detail;
+			return null;
+		});
+
+	if (error) {
+		throw error;
+	}
+
+	return res;
+};
+
 export const userSignIn = async (email: string, password: string) => {
 	let error = null;
 
@@ -17,10 +44,10 @@ export const userSignIn = async (email: string, password: string) => {
 			if (!res.ok) throw await res.json();
 			return res.json();
 		})
-		.catch((error) => {
-			console.log(error);
+		.catch((err) => {
+			console.log(err);
 
-			error = error.detail;
+			error = err.detail;
 			return null;
 		});
 
@@ -49,9 +76,9 @@ export const userSignUp = async (name: string, email: string, password: string) 
 			if (!res.ok) throw await res.json();
 			return res.json();
 		})
-		.catch((error) => {
-			console.log(error);
-			error = error.detail;
+		.catch((err) => {
+			console.log(err);
+			error = err.detail;
 			return null;
 		});
